Rename cards and players
